Split decoding and encoding tests.
